+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 ## unreleased (2023-mm-dd)
 ### Added
 ### Changed
+### Fixed
+### Deprecated
+### Removed
+### Security
+## 0.0.4 (2023-12-07)
+### Added
+- redefine openapi for application/xml response type
+### Changed
 - Logging to ipytv extended
+- update README.md
 ### Fixed
 ### Deprecated
 ### Removed